Move FixFollowEvents to workflow handler
